CB-6184: Fix infer_shape of silu in ArmActivationOp
torch.nn.SiLU doesn't support float16 on cpu.

+++ AIPUBuilder/Parser/ops/release_ops.py
@@ -179,7 +179,7 @@ class ArmActivationOp(LayoutUnawareOp, OpHasMethod, OpHasOneOutPort, ArmOp):
 
     def silu(self):
         inputs = self.get_input_tensors()
-        tor_arr = torch.from_numpy(inputs[0])
+        tor_arr = torch.from_numpy(inputs[0].astype(np.float32))
         m = torch.nn.SiLU()
         out_tensor = m(tor_arr)
         tor2numpy = out_tensor.numpy().astype(inputs[0].dtype)
